Add contents of http://d-lev.com/support/d-lib_2023-11-02.zip

+++ go/d-lib_time.go
@@ -28,6 +28,6 @@ func date_hm() (string) {
 }
 
 // return time for use as random seed
-func timeseed() (int) {
-	return int(time.Now().UnixNano())
+func timeseed() (int64) {
+	return time.Now().UnixNano()
 }
